refactor: 프로젝트 코드 파일 분리
